Adding the concept of the "user agent header".
The user agent header is a DocumentHeader that is added to all documents,
before any parsing occurs. (It is also added to empty documents that are
created.) This allows scripts and stylesheets to be loaded in all documents.
Note that this functionality isn't very thoroughly tested, but at least works
with inline stylesheets.